requirements.txt
selenium
pytest
allure-pytest
pytest-order